Added 1 new ECR check. Added VPC flow log playbook. Updated CFN, TF and standalone policies. Updated Readmes.
Added a new check to ECR that checks if the latest pushed image in every ECR repo has any vulnerabilities and fails the check if they do regardless of the CVSS scoring. Added a new VPC flow logging playbook, makes use of the `raise` functionality in Python to exit the remediation sooner if there is an existing resource. Added a zipped version and created an entry in the `Directors Cut` of the CFN template. Aligned all IAM permissions for ElectricEye-Response in CFN and TF, they were not matching. Updated the ServiceNow architecture diagram. Updated readmes. Started swapping Auditors to use `Severity.Label` in the ASFF