Add a Pagerduty integration add-on. Updated Readmes for additional tools and bad grammar.
Added a new ElectricEye module titled `electriceye-pagerduty-integration`. This was supposed to be part of ElectricEye-ChatOps, but it's not a Chat tool per se so it is in its own section. It probably belonged in ElectricEye-Response in retrospect. Added new readmes, TF, CFN and screenshots to support this. Added new tools, grammar fixes and known issue entries to the core module Readme. Added a TBD for Microsoft Teams integration.