Add AWS Glue auditor with 6 checks. Update CFN, TF and standalone policies. Update readme
Added an AWS Glue auditor with 6 checks, 3 for crawlers and 3 for the data catalog - all except 1 are encryption related with the last being a resource-based policy check for the catalog. Updated the permissions in the CFN, TF and standalone policies. Updated the readme. New count is 201 checks, 60 services/components and 43 auditors.