Add KDF and IMDSv2 Auditors. Added checks to SecSvc and SG Auditors. Update permissions in CF, TF and Standalone. Update readme.
Add a Kinesis Data Firehose Auditor with 1 check, an EC2 Auditor (non-SSM) with 1 check and added 1 new check to Security Services (detective) and 2 new checks to the SG Auditor (Kafka and Cassandra open). Updated IAM permissions in CFN, TF and the standalone policy. Updated the readme and shell script for the auditors. Total checks are 206 for 62 services with 45 auditors.